+++ src/mame/tatsumi/tatsumi_m.cpp
@@ -18,108 +18,6 @@ void tatsumi_state::tatsumi_reset()
 
 /******************************************************************************/
 
-uint16_t apache3_state::apache3_bank_r()
-{
-	return m_control_word;
-}
-
-void apache3_state::apache3_bank_w(offs_t offset, uint16_t data, uint16_t mem_mask)
-{
-	/*
-	    0x8000  - Set when accessing palette ram (not implemented, perhaps blank screen?)
-	    0x0080  - Set when accessing IO cpu RAM/ROM (implemented as halt cpu)
-	    0x0060  - IOP bank to access from main cpu (0x0 = RAM, 0x20 = lower ROM, 0x60 = upper ROM)
-	    0x0010  - Set when accessing OBJ cpu RAM/ROM (implemented as halt cpu)
-	    0x000f  - OBJ bank to access from main cpu (0x8 = RAM, 0x0 to 0x7 = ROM)
-	*/
-
-	COMBINE_DATA(&m_control_word);
-
-	if (m_control_word & 0x7f00)
-	{
-		logerror("Unknown control Word: %04x\n",m_control_word);
-		m_subcpu2->set_input_line(INPUT_LINE_HALT, CLEAR_LINE); // ?
-	}
-
-	if (m_control_word & 0x10)
-		m_subcpu->set_input_line(INPUT_LINE_HALT, ASSERT_LINE);
-	else
-		m_subcpu->set_input_line(INPUT_LINE_HALT, CLEAR_LINE);
-
-	if (m_control_word & 0x80)
-		m_audiocpu->set_input_line(INPUT_LINE_HALT, ASSERT_LINE);
-	else
-		m_audiocpu->set_input_line(INPUT_LINE_HALT, CLEAR_LINE);
-
-	m_last_control=m_control_word;
-}
-
-// D1 = /ZBREQ  - Z80 bus request
-// D0 = /GRDACC - Allow 68000 access to road pattern RAM
-void apache3_state::apache3_z80_ctrl_w(uint16_t data)
-{
-	m_subcpu2->set_input_line(INPUT_LINE_HALT, data & 2 ? ASSERT_LINE : CLEAR_LINE);
-}
-
-uint16_t apache3_state::apache3_v30_v20_r(offs_t offset)
-{
-	address_space &targetspace = m_audiocpu->space(AS_PROGRAM);
-
-	/* Each V20 byte maps to a V30 word */
-	if ((m_control_word & 0xe0) == 0xe0)
-		offset += 0xf8000; /* Upper half */
-	else if ((m_control_word & 0xe0) == 0xc0)
-		offset += 0xf0000;
-	else if ((m_control_word & 0xe0) == 0x80)
-		offset += 0x00000; // main ram
-	else
-		logerror("%08x: unmapped read z80 rom %08x\n", m_maincpu->pc(), offset);
-	return 0xff00 | targetspace.read_byte(offset);
-}
-
-void apache3_state::apache3_v30_v20_w(offs_t offset, uint16_t data, uint16_t mem_mask)
-{
-	address_space &targetspace = m_audiocpu->space(AS_PROGRAM);
-
-	if ((m_control_word & 0xe0) != 0x80)
-		logerror("%08x: write unmapped v30 rom %08x\n", m_maincpu->pc(), offset);
-
-	/* Only 8 bits of the V30 data bus are connected - ignore writes to the other half */
-	if (ACCESSING_BITS_0_7)
-	{
-		targetspace.write_byte(offset, data & 0xff);
-	}
-}
-
-uint16_t apache3_state::apache3_z80_r(offs_t offset)
-{
-	return m_apache3_z80_ram[offset];
-}
-
-void apache3_state::apache3_z80_w(offs_t offset, uint16_t data)
-{
-	m_apache3_z80_ram[offset] = data & 0xff;
-}
-
-uint8_t apache3_state::apache3_vr1_r()
-{
-	return (uint8_t)((255./100) * (100 - m_vr1->read()));
-}
-
-/* Ground/sky rotation control
- *
- * There are 12 16-bit values that are
- * presumably loaded into the 8 TZ2213 custom
- * accumulators and counters.
- */
-void apache3_state::apache3_rotate_w(uint16_t data)
-{
-	m_apache3_rotate_ctrl[m_apache3_rot_idx] = data;
-	m_apache3_rot_idx = (m_apache3_rot_idx + 1) % 12;
-}
-
-/******************************************************************************/
-
 uint16_t roundup5_state::roundup_v30_z80_r(offs_t offset)
 {
 	address_space &targetspace = m_audiocpu->space(AS_PROGRAM);
