+++ src/mame/tatsumi/tatsumi_v.cpp
@@ -92,126 +92,7 @@ TILE_GET_INFO_MEMBER(tatsumi_state::get_text_tile_info)
 }
 
 
-/**********************************
- *
- * Apache 3
- *
- *********************************/
 
-void apache3_state::apache3_road_z_w(uint16_t data)
-{
-	m_apache3_road_z = data & 0xff;
-}
-
-void apache3_state::apache3_road_x_w(offs_t offset, uint8_t data)
-{
-	// Note: Double buffered. Yes, this is correct :)
-	m_apache3_road_x_ram[data] = offset;
-}
-
-void apache3_state::draw_sky(bitmap_rgb32 &bitmap,const rectangle &cliprect, int palette_base, int start_offset)
-{
-	// all TODO
-	if (start_offset&0x8000)
-		start_offset=-(0x10000 - start_offset);
-
-	start_offset=-start_offset;
-
-	start_offset-=48;
-	for (int y=0; y<256; y++) {
-		for (int x=0; x<320; x++) {
-			int col=palette_base + y + start_offset;
-			if (col<palette_base) col=palette_base;
-			if (col>palette_base+127) col=palette_base+127;
-
-			bitmap.pix(y, x) = m_palette->pen(col);
-		}
-	}
-}
-
-/* Draw the sky and ground, applying rotation (eventually). Experimental! */
-void apache3_state::draw_ground(bitmap_rgb32 &dst, const rectangle &cliprect)
-{
-	if (0)
-	{
-		uint16_t gva = 0x180; // TODO
-		uint8_t sky_val = m_apache3_rotate_ctrl[1] & 0xff;
-
-		for (int y = cliprect.min_y; y <= cliprect.max_y; ++y)
-		{
-			uint16_t rgdb = 0;//m_apache3_road_x_ram[gva & 0xff];
-			uint16_t gha = 0xf60; // test
-			int ln = (((m_apache3_prom[gva & 0x7f] & 0x7f) + (m_apache3_road_z & 0x7f)) >> 5) & 3;
-
-			if (gva & 0x100)
-			{
-				/* Sky */
-				for (int x = cliprect.min_x; x <= cliprect.max_x; ++x)
-				{
-					dst.pix(y, x) = m_palette->pen(0x100 + (sky_val & 0x7f));
-
-					/* Update horizontal counter? */
-					gha = (gha + 1) & 0xfff;
-				}
-			}
-			else
-			{
-				/* Ground */
-				for (int x = cliprect.min_x; x <= cliprect.max_x; ++x)
-				{
-					uint16_t hval = (rgdb + gha) & 0xfff; // Not quite
-
-					if (hval & 0x800)
-						hval ^= 0x1ff; // TEST
-					//else
-						//hval = hval;
-
-					uint8_t pixels = m_apache3_g_ram[(((gva & 0xff) << 7) | ((hval >> 2) & 0x7f))];
-					int pix_sel = hval & 3;
-
-					uint8_t colour = (pixels >> (pix_sel << 1)) & 3;
-					colour = (BIT(hval, 11) << 4) | (colour << 2) | ln;
-
-					/* Draw the pixel */
-					dst.pix(y, x) = m_palette->pen(0x200 + colour);
-
-					/* Update horizontal counter */
-					gha = (gha + 1) & 0xfff;
-				}
-			}
-
-			/* Update sky counter */
-			sky_val++;
-			gva = (gva + 1) & 0x1ff;
-		}
-	}
-}
-
-
-VIDEO_START_MEMBER(apache3_state,apache3)
-{
-	m_tx_layer = &machine().tilemap().create(*m_gfxdecode, tilemap_get_info_delegate(*this, FUNC(tatsumi_state::get_text_tile_info)), TILEMAP_SCAN_ROWS, 8,8, 64,64);
-	m_apache3_road_x_ram = std::make_unique<uint8_t[]>(512);
-
-	m_tx_layer->set_transparent_pen(0);
-}
-
-uint32_t apache3_state::screen_update_apache3(screen_device &screen, bitmap_rgb32 &bitmap, const rectangle &cliprect)
-{
-	m_rotatingsprites->update_cluts();
-
-	m_tx_layer->set_scrollx(0,24);
-
-	bitmap.fill(m_palette->pen(0), cliprect);
-	screen.priority().fill(0, cliprect);
-	m_rotatingsprites->draw_sprites(screen.priority(),cliprect,1,(m_sprite_control_ram[0xe0]&0x1000) ? 0x1000 : 0); // Alpha pass only
-	draw_sky(bitmap, cliprect, 256, m_apache3_rotate_ctrl[1]);
-	apply_shadow_bitmap(bitmap,cliprect,screen.priority(), 0);
-//  draw_ground(bitmap, cliprect);
-	m_rotatingsprites->draw_sprites(bitmap,cliprect,0, (m_sprite_control_ram[0x20]&0x1000) ? 0x1000 : 0);
-	m_tx_layer->draw(screen, bitmap, cliprect, 0,0);
-	return 0;
-}
 
 /**********************************
  *
